pyproject.toml
[project]
name = "youtube-vision-transcriber"
version = "0.1.0"
description = "AI-powered pipeline that converts YouTube videos into polished articles using vision-based transcription"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "fast-agent-mcp==0.4.31",
    "fastmcp==2.13.1",
    "yt-dlp==2025.11.12",
]

[project.scripts]
youtube-vision-transcriber = "agent:start_server"


